refactor: extract CoreObjectPickerHelper to eliminate duplicated type/reference picker logic
