+++ src/Beutl/Views/Editors/CoreObjectEditor.axaml.cs
@@ -114,17 +114,7 @@ public partial class CoreObjectEditor : UserControl
         try
         {
             _flyoutOpen = true;
-            Type propertyType = viewModel.PropertyAdapter.PropertyType;
-            string format = propertyType.FullName!;
-            var selectVm = new SelectLibraryItemDialogViewModel(format, propertyType);
-
-            if (PresenterTypeAttribute.GetPresenterType(propertyType) != null)
-            {
-                var targets = viewModel.GetAvailableTargets();
-                selectVm.InitializeReferences(targets);
-            }
-
-            return await LibraryItemPickerHelper.ShowAsync(this, selectVm, format);
+            return await CoreObjectPickerHelper.ShowTypeOrReferenceAsync(this, viewModel);
         }
         finally
         {
